+++ src/cli/commands/anon-col/anon-col/command.ts
@@ -1,6 +1,5 @@
-import {Command, Option} from 'commander';
-import {providers} from '../../../../types/types';
-import {maskCommand} from '../commands/mask/mask-command';
+import {Command} from 'commander';
+import {maskCommand} from '../commands/mask/command';
 
 export const anonColCommand = new Command('anon-col');
 
@@ -11,10 +10,4 @@ anonColCommand
 	.argument('<table>', 'table name')
 	.argument('<column>', 'column name')
 	.addCommand(maskCommand, {isDefault: true})
-	.addOption(
-		new Option('-p --provider <provider>', 'provider to be used for column')
-			.default('mask')
-			.choices(providers),
-	)
-	.option('-skip --no-confirm', 'Do not wait for user confirmation');
-// .action(anonColAction);
+	.option('-C --no-confirm', 'Do not wait for user confirmation');
